requirements.txt
playwright
python-dotenv
groq
gspread
filelock
